requirements.txt
dash==2.1.0
dash_bootstrap_components==1.0.3
flask==2.0.2
gunicorn==20.1.0
matplotlib==3.4.3
numpy>=1.22
pandas==1.3.5
Pillow==9.1.1
plotly==5.5.0
requests==2.24.0
torch==1.10.2
torchvision==0.11.3
werkzeug==2.0.3
pathlib
